feat: introduce permission manager
Moved group methods to group package
Moved createTestControllerEnvironment function to jimmtest package
