+++ internal/jimm/cloud_test.go
@@ -975,473 +975,6 @@ func TestAddCloudToController(t *testing.T) {
 	}
 }
 
-const grantCloudAccessTestEnv = `clouds:
-- name: test-cloud
-  type: test-provider
-  regions:
-  - name: test-cloud-region
-- name: test
-  type: kubernetes
-  host-cloud-region: test-cloud/test-cloud-region
-  regions:
-  - name: default
-  - name: region2
-  users:
-  - user: alice@canonical.com
-    access: admin
-controllers:
-- name: controller-1
-  uuid: 00000001-0000-0000-0000-000000000001
-  cloud: test-cloud
-  region: test-cloud-region
-  cloud-regions:
-  - cloud: test-cloud
-    region: test-cloud-region
-    priority: 10
-  - cloud: test
-    region: default
-    priority: 1
-  - cloud: test
-    region: region2
-    priority: 1
-`
-
-var grantCloudAccessTests = []struct {
-	name            string
-	env             string
-	dialError       error
-	username        string
-	cloud           string
-	targetUsername  string
-	access          string
-	expectRelations []openfga.Tuple
-	expectError     string
-	expectErrorCode errors.Code
-}{{
-	name:            "CloudNotFound",
-	username:        "alice@canonical.com",
-	cloud:           "test2",
-	targetUsername:  "bob@canonical.com",
-	access:          "add-model",
-	expectError:     `cloud "test2" not found`,
-	expectErrorCode: errors.CodeNotFound,
-}, {
-	name:           "Admin grants admin access",
-	env:            grantCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "bob@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-}, {
-	name:           "Admin grants add-model access",
-	env:            grantCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "bob@canonical.com",
-	access:         "add-model",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.CanAddModelRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-}, {
-	name:            "UserNotAuthorized",
-	env:             grantCloudAccessTestEnv,
-	username:        "charlie@canonical.com",
-	cloud:           "test",
-	targetUsername:  "bob@canonical.com",
-	access:          "add-model",
-	expectError:     `unauthorized`,
-	expectErrorCode: errors.CodeUnauthorized,
-}, {
-	name:           "DialError",
-	env:            grantCloudAccessTestEnv,
-	dialError:      errors.E("test dial error"),
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "bob@canonical.com",
-	access:         "add-model",
-	expectError:    `test dial error`,
-}, {
-	name:           "unknown access",
-	env:            grantCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "bob@canonical.com",
-	access:         "some-unknown-access",
-	expectError:    `failed to recognize given access: "some-unknown-access"`,
-}}
-
-func TestGrantCloudAccess(t *testing.T) {
-	c := qt.New(t)
-
-	for _, t := range grantCloudAccessTests {
-		tt := t
-		c.Run(tt.name, func(c *qt.C) {
-			ctx := context.Background()
-
-			env := jimmtest.ParseEnvironment(c, tt.env)
-			dialer := &jimmtest.Dialer{
-				API: &jimmtest.API{},
-				Err: tt.dialError,
-			}
-
-			j := jimmtest.NewJIMM(c, &jimm.Parameters{
-				Dialer: dialer,
-			})
-
-			env.PopulateDBAndPermissions(c, j.ResourceTag(), j.Database, j.OpenFGAClient)
-
-			dbUser := env.User(tt.username).DBObject(c, j.Database)
-			user := openfga.NewUser(&dbUser, j.OpenFGAClient)
-
-			err := j.GrantCloudAccess(ctx, user, names.NewCloudTag(tt.cloud), names.NewUserTag(tt.targetUsername), tt.access)
-			c.Assert(dialer.IsClosed(), qt.Equals, true)
-			if tt.expectError != "" {
-				c.Check(err, qt.ErrorMatches, tt.expectError)
-				if tt.expectErrorCode != "" {
-					c.Check(errors.ErrorCode(err), qt.Equals, tt.expectErrorCode)
-				}
-				return
-			}
-			c.Assert(err, qt.IsNil)
-			for _, tuple := range tt.expectRelations {
-				value, err := j.OpenFGAClient.CheckRelation(ctx, tuple, false)
-				c.Assert(err, qt.IsNil)
-				c.Assert(value, qt.IsTrue, qt.Commentf("expected the tuple to exist after granting"))
-			}
-		})
-	}
-}
-
-const revokeCloudAccessTestEnv = `clouds:
-- name: test-cloud
-  type: test-provider
-  regions:
-  - name: test-cloud-region
-  users:
-  - user: daphne@canonical.com
-    access: admin
-- name: test
-  type: kubernetes
-  host-cloud-region: test-cloud/test-cloud-region
-  regions:
-  - name: default
-  users:
-  - user: alice@canonical.com
-    access: admin
-  - user: bob@canonical.com
-    access: admin
-  - user: charlie@canonical.com
-    access: add-model
-controllers:
-- name: controller-1
-  uuid: 00000001-0000-0000-0000-000000000001
-  cloud: test-cloud
-  region: test-cloud-region
-  cloud-regions:
-  - cloud: test-cloud
-    region: test-cloud-region
-    priority: 10
-  - cloud: test
-    region: default
-    priority: 1
-`
-
-var revokeCloudAccessTests = []struct {
-	name                   string
-	env                    string
-	dialError              error
-	username               string
-	cloud                  string
-	targetUsername         string
-	access                 string
-	extraInitialTuples     []openfga.Tuple
-	expectRelations        []openfga.Tuple
-	expectRemovedRelations []openfga.Tuple
-	expectError            string
-	expectErrorCode        errors.Code
-}{{
-	name:            "CloudNotFound",
-	username:        "alice@canonical.com",
-	cloud:           "test2",
-	targetUsername:  "bob@canonical.com",
-	access:          "admin",
-	expectError:     `cloud "test2" not found`,
-	expectErrorCode: errors.CodeNotFound,
-}, {
-	name:           "Admin revokes 'admin' from another admin",
-	env:            revokeCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "bob@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.CanAddModelRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-}, {
-	name:           "Admin revokes 'add-model' from another admin",
-	env:            revokeCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "bob@canonical.com",
-	access:         "add-model",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.CanAddModelRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-}, {
-	name:           "Admin revokes 'add-model' from a user with 'add-model' access",
-	env:            revokeCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "charlie@canonical.com",
-	access:         "add-model",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.CanAddModelRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-}, {
-	name:           "Admin revokes 'add-model' from a user with no access",
-	env:            revokeCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "daphne@canonical.com",
-	access:         "add-model",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.CanAddModelRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-}, {
-	name:           "Admin revokes 'admin' from a user with no access",
-	env:            revokeCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "daphne@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.CanAddModelRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-}, {
-	name:           "Admin revokes 'add-model' access from a user who has separate tuples for all accesses (add-model/admin)",
-	env:            revokeCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "charlie@canonical.com",
-	access:         "add-model",
-	extraInitialTuples: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	},
-	// No need to add the 'add-model' relation, because it's already there due to the environment setup.
-	},
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.CanAddModelRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-}, {
-	name:           "Admin revokes 'admin' access from a user who has separate tuples for all accesses (add-model/admin)",
-	env:            revokeCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "charlie@canonical.com",
-	access:         "admin",
-	extraInitialTuples: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	},
-	// No need to add the 'add-model' relation, because it's already there due to the environment setup.
-	},
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.CanAddModelRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewCloudTag("test")),
-	}},
-}, {
-	name:            "UserNotAuthorized",
-	env:             revokeCloudAccessTestEnv,
-	username:        "charlie@canonical.com",
-	cloud:           "test",
-	targetUsername:  "bob@canonical.com",
-	access:          "add-model",
-	expectError:     `unauthorized`,
-	expectErrorCode: errors.CodeUnauthorized,
-}, {
-	name:           "DialError",
-	env:            revokeCloudAccessTestEnv,
-	dialError:      errors.E("test dial error"),
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "bob@canonical.com",
-	access:         "add-model",
-	expectError:    `test dial error`,
-}, {
-	name:           "unknown access",
-	env:            revokeCloudAccessTestEnv,
-	username:       "alice@canonical.com",
-	cloud:          "test",
-	targetUsername: "bob@canonical.com",
-	access:         "some-unknown-access",
-	expectError:    `failed to recognize given access: "some-unknown-access"`,
-}}
-
-//nolint:gocognit
-func TestRevokeCloudAccess(t *testing.T) {
-	c := qt.New(t)
-
-	for _, t := range revokeCloudAccessTests {
-		tt := t
-		c.Run(tt.name, func(c *qt.C) {
-			ctx := context.Background()
-
-			env := jimmtest.ParseEnvironment(c, tt.env)
-			dialer := &jimmtest.Dialer{
-				API: &jimmtest.API{},
-				Err: tt.dialError,
-			}
-
-			j := jimmtest.NewJIMM(c, &jimm.Parameters{
-				Dialer: dialer,
-			})
-
-			env.PopulateDBAndPermissions(c, j.ResourceTag(), j.Database, j.OpenFGAClient)
-
-			if len(tt.extraInitialTuples) > 0 {
-				err := j.OpenFGAClient.AddRelation(ctx, tt.extraInitialTuples...)
-				c.Assert(err, qt.IsNil)
-			}
-
-			if tt.expectRemovedRelations != nil {
-				for _, tuple := range tt.expectRemovedRelations {
-					value, err := j.OpenFGAClient.CheckRelation(ctx, tuple, false)
-					c.Assert(err, qt.IsNil)
-					c.Assert(value, qt.IsTrue, qt.Commentf("expected the tuple to exist before revoking"))
-				}
-			}
-
-			dbUser := env.User(tt.username).DBObject(c, j.Database)
-			user := openfga.NewUser(&dbUser, j.OpenFGAClient)
-
-			err := j.RevokeCloudAccess(ctx, user, names.NewCloudTag(tt.cloud), names.NewUserTag(tt.targetUsername), tt.access)
-			c.Assert(dialer.IsClosed(), qt.Equals, true)
-			if tt.expectError != "" {
-				c.Check(err, qt.ErrorMatches, tt.expectError)
-				if tt.expectErrorCode != "" {
-					c.Check(errors.ErrorCode(err), qt.Equals, tt.expectErrorCode)
-				}
-				return
-			}
-			c.Assert(err, qt.IsNil)
-			if tt.expectRemovedRelations != nil {
-				for _, tuple := range tt.expectRemovedRelations {
-					value, err := j.OpenFGAClient.CheckRelation(ctx, tuple, false)
-					c.Assert(err, qt.IsNil)
-					c.Assert(value, qt.IsFalse, qt.Commentf("expected the tuple to be removed after revoking"))
-				}
-			}
-			if tt.expectRelations != nil {
-				for _, tuple := range tt.expectRelations {
-					value, err := j.OpenFGAClient.CheckRelation(ctx, tuple, false)
-					c.Assert(err, qt.IsNil)
-					c.Assert(value, qt.IsTrue, qt.Commentf("expected the tuple to exist after revoking"))
-				}
-			}
-		})
-	}
-}
-
 const removeCloudTestEnv = `clouds:
 - name: test-cloud
   type: test-provider
