+++ internal/jimm/export_test.go
@@ -11,17 +11,14 @@ import (
 	"github.com/canonical/jimm/v3/internal/db"
 	"github.com/canonical/jimm/v3/internal/dbmodel"
 	"github.com/canonical/jimm/v3/internal/openfga"
-	ofganames "github.com/canonical/jimm/v3/internal/openfga/names"
 )
 
 var (
-	DetermineAccessLevelAfterGrant = determineAccessLevelAfterGrant
-	PollDuration                   = pollDuration
-	CalculateNextPollDuration      = calculateNextPollDuration
-	NewControllerClient            = &newControllerClient
-	FillMigrationTarget            = fillMigrationTarget
-	InitiateMigration              = &initiateMigration
-	ResolveTag                     = resolveTag
+	PollDuration              = pollDuration
+	CalculateNextPollDuration = calculateNextPollDuration
+	NewControllerClient       = &newControllerClient
+	FillMigrationTarget       = fillMigrationTarget
+	InitiateMigration         = &initiateMigration
 )
 
 func NewWatcherWithControllerUnavailableChan(db *db.Database, dialer Dialer, pubsub Publisher, testChannel chan error) *Watcher {
@@ -44,10 +41,6 @@ func NewWatcherWithDeltaProcessedChannel(db *db.Database, dialer Dialer, pubsub 
 
 func (j *JIMM) ListApplicationOfferUsers(ctx context.Context, offer names.ApplicationOfferTag, user *dbmodel.Identity, adminAccess bool) ([]jujuparams.OfferUserDetails, error) {
 	return j.listApplicationOfferUsers(ctx, offer, user, adminAccess)
-}
-
-func (j *JIMM) ParseAndValidateTag(ctx context.Context, key string) (*ofganames.Tag, error) {
-	return j.parseAndValidateTag(ctx, key)
 }
 
 func (j *JIMM) EveryoneUser() *openfga.User {
