+++ internal/jimm/model_test.go
@@ -26,7 +26,6 @@ import (
 	"github.com/canonical/jimm/v3/internal/errors"
 	"github.com/canonical/jimm/v3/internal/jimm"
 	"github.com/canonical/jimm/v3/internal/openfga"
-	ofganames "github.com/canonical/jimm/v3/internal/openfga/names"
 	"github.com/canonical/jimm/v3/internal/testutils/jimmtest"
 )
 
@@ -1999,997 +1998,6 @@ func TestModelSummaries(t *testing.T) {
 			c.Check(err, qt.IsNil)
 			c.Check(summaries.Results, qt.HasLen, t.expectedSummariesSize)
 			c.Check(summaries.Results, qt.DeepEquals, t.expectedSummaries)
-		})
-	}
-}
-
-const grantModelAccessTestEnv = `clouds:
-- name: test-cloud
-  type: test-provider
-  regions:
-  - name: test-cloud-region
-cloud-credentials:
-- owner: alice@canonical.com
-  name: cred-1
-  cloud: test-cloud
-controllers:
-- name: controller-1
-  uuid: 00000001-0000-0000-0000-000000000001
-  cloud: test-cloud
-  region: test-cloud-region
-models:
-- name: model-1
-  uuid: 00000002-0000-0000-0000-000000000001
-  controller: controller-1
-  cloud: test-cloud
-  region: test-cloud-region
-  cloud-credential: cred-1
-  owner: alice@canonical.com
-  users:
-  - user: alice@canonical.com
-    access: admin
-  - user: charlie@canonical.com
-    access: write
-`
-
-var grantModelAccessTests = []struct {
-	name            string
-	env             string
-	dialError       error
-	username        string
-	uuid            string
-	targetUsername  string
-	access          string
-	expectRelations []openfga.Tuple
-	expectError     string
-	expectErrorCode errors.Code
-}{{
-	name:            "ModelNotFound",
-	username:        "alice@canonical.com",
-	uuid:            "00000002-0000-0000-0000-000000000001",
-	targetUsername:  "bob@canonical.com",
-	access:          "write",
-	expectError:     `model not found`,
-	expectErrorCode: errors.CodeNotFound,
-}, {
-	name:           "Admin grants 'admin' access to a user with no access",
-	env:            grantModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "bob@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin grants 'write' access to a user with no access",
-	env:            grantModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "bob@canonical.com",
-	access:         "write",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin grants 'read' access to a user with no access",
-	env:            grantModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "bob@canonical.com",
-	access:         "read",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin grants 'write' access to a user who already has 'write' access",
-	env:            grantModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "charlie@canonical.com",
-	access:         "write",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin grants 'read' access to a user who already has 'write' access",
-	env:            grantModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "charlie@canonical.com",
-	access:         "read",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin grants 'admin' access to themselves",
-	env:            grantModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "alice@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin grants 'write' access to themselves",
-	env:            grantModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "alice@canonical.com",
-	access:         "write",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin grants 'read' access to themselves",
-	env:            grantModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "alice@canonical.com",
-	access:         "read",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:            "UserNotAuthorized",
-	env:             grantModelAccessTestEnv,
-	username:        "charlie@canonical.com",
-	uuid:            "00000002-0000-0000-0000-000000000001",
-	targetUsername:  "bob@canonical.com",
-	access:          "write",
-	expectError:     `unauthorized`,
-	expectErrorCode: errors.CodeUnauthorized,
-}, {
-	name:           "DialError",
-	env:            grantModelAccessTestEnv,
-	dialError:      errors.E("test dial error"),
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "bob@canonical.com",
-	access:         "write",
-	expectError:    `test dial error`,
-}, {
-	name:           "unknown access",
-	env:            grantModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "bob@canonical.com",
-	access:         "some-unknown-access",
-	expectError:    `failed to recognize given access: "some-unknown-access"`,
-}}
-
-func TestGrantModelAccess(t *testing.T) {
-	c := qt.New(t)
-
-	for _, t := range grantModelAccessTests {
-		tt := t
-		c.Run(tt.name, func(c *qt.C) {
-			ctx := context.Background()
-
-			dialer := &jimmtest.Dialer{
-				API: &jimmtest.API{},
-				Err: tt.dialError,
-			}
-			j := jimmtest.NewJIMM(c, &jimm.Parameters{
-				Dialer: dialer,
-			})
-
-			env := jimmtest.ParseEnvironment(c, tt.env)
-			env.PopulateDBAndPermissions(c, j.ResourceTag(), j.Database, j.OpenFGAClient)
-
-			dbUser := env.User(tt.username).DBObject(c, j.Database)
-			user := openfga.NewUser(&dbUser, j.OpenFGAClient)
-
-			err := j.GrantModelAccess(ctx, user, names.NewModelTag(tt.uuid), names.NewUserTag(tt.targetUsername), jujuparams.UserAccessPermission(tt.access))
-			c.Assert(dialer.IsClosed(), qt.IsTrue)
-			if tt.expectError != "" {
-				c.Check(err, qt.ErrorMatches, tt.expectError)
-				if tt.expectErrorCode != "" {
-					c.Check(errors.ErrorCode(err), qt.Equals, tt.expectErrorCode)
-				}
-				return
-			}
-			c.Assert(err, qt.IsNil)
-			for _, tuple := range tt.expectRelations {
-				value, err := j.OpenFGAClient.CheckRelation(ctx, tuple, false)
-				c.Assert(err, qt.IsNil)
-				c.Assert(value, qt.IsTrue, qt.Commentf("expected the tuple to exist after granting"))
-			}
-		})
-	}
-}
-
-const revokeModelAccessTestEnv = `clouds:
-- name: test-cloud
-  type: test-provider
-  regions:
-  - name: test-cloud-region
-cloud-credentials:
-- owner: alice@canonical.com
-  name: cred-1
-  cloud: test-cloud
-controllers:
-- name: controller-1
-  uuid: 00000001-0000-0000-0000-000000000001
-  cloud: test-cloud
-  region: test-cloud-region
-models:
-- name: model-1
-  uuid: 00000002-0000-0000-0000-000000000001
-  controller: controller-1
-  cloud: test-cloud
-  region: test-cloud-region
-  cloud-credential: cred-1
-  owner: alice@canonical.com
-  users:
-  - user: alice@canonical.com
-    access: admin
-  - user: bob@canonical.com
-    access: admin
-  - user: charlie@canonical.com
-    access: write
-  - user: daphne@canonical.com
-    access: read
-- name: model-2
-  uuid: 00000002-0000-0000-0000-000000000002
-  controller: controller-1
-  cloud: test-cloud
-  region: test-cloud-region
-  cloud-credential: cred-1
-  owner: alice@canonical.com
-  users:
-  - user: alice@canonical.com
-    access: admin
-  - user: earl@canonical.com
-    access: admin
-`
-
-var revokeModelAccessTests = []struct {
-	name                   string
-	env                    string
-	dialError              error
-	username               string
-	uuid                   string
-	targetUsername         string
-	access                 string
-	extraInitialTuples     []openfga.Tuple
-	expectRelations        []openfga.Tuple
-	expectRemovedRelations []openfga.Tuple
-	expectError            string
-	expectErrorCode        errors.Code
-}{{
-	name:            "ModelNotFound",
-	username:        "alice@canonical.com",
-	uuid:            "00000002-0000-0000-0000-000000000001",
-	targetUsername:  "bob@canonical.com",
-	access:          "write",
-	expectError:     `model not found`,
-	expectErrorCode: errors.CodeNotFound,
-}, {
-	name:           "Admin revokes 'admin' access from another admin",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "bob@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'write' access from another admin",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "bob@canonical.com",
-	access:         "write",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'read' access from another admin",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "bob@canonical.com",
-	access:         "read",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'admin' access from a user who has 'write' access",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "charlie@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'write' access from a user who has 'write' access",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "charlie@canonical.com",
-	access:         "write",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'read' access from a user who has 'write' access",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "charlie@canonical.com",
-	access:         "read",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'admin' access from a user who has 'read' access",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "daphne@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'write' access from a user who has 'read' access",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "daphne@canonical.com",
-	access:         "write",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'read' access from a user who has 'read' access",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "daphne@canonical.com",
-	access:         "read",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'admin' access from themselves",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "alice@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'write' access from themselves",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "alice@canonical.com",
-	access:         "write",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'read' access from themselves",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "alice@canonical.com",
-	access:         "read",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Writer revokes 'admin' access from themselves",
-	env:            revokeModelAccessTestEnv,
-	username:       "charlie@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "charlie@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Writer revokes 'write' access from themselves",
-	env:            revokeModelAccessTestEnv,
-	username:       "charlie@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "charlie@canonical.com",
-	access:         "write",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Writer revokes 'read' access from themselves",
-	env:            revokeModelAccessTestEnv,
-	username:       "charlie@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "charlie@canonical.com",
-	access:         "read",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Reader revokes 'admin' access from themselves",
-	env:            revokeModelAccessTestEnv,
-	username:       "daphne@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "daphne@canonical.com",
-	access:         "admin",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Reader revokes 'write' access from themselves",
-	env:            revokeModelAccessTestEnv,
-	username:       "daphne@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "daphne@canonical.com",
-	access:         "write",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Reader revokes 'read' access from themselves",
-	env:            revokeModelAccessTestEnv,
-	username:       "daphne@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "daphne@canonical.com",
-	access:         "read",
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'admin' access from a user who has separate tuples for all accesses (read/write/admin)",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "daphne@canonical.com",
-	access:         "admin",
-	extraInitialTuples: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	},
-	// No need to add the 'read' relation, because it's already there due to the environment setup.
-	},
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'write' access from a user who has separate tuples for all accesses (read/write/admin)",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "daphne@canonical.com",
-	access:         "write",
-	extraInitialTuples: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	},
-	// No need to add the 'read' relation, because it's already there due to the environment setup.
-	},
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:           "Admin revokes 'read' access from a user who has separate tuples for all accesses (read/write/admin)",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "daphne@canonical.com",
-	access:         "read",
-	extraInitialTuples: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	},
-	// No need to add the 'read' relation, because it's already there due to the environment setup.
-	},
-	expectRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("alice@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("bob@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("charlie@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-	expectRemovedRelations: []openfga.Tuple{{
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.AdministratorRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.WriterRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}, {
-		Object:   ofganames.ConvertTag(names.NewUserTag("daphne@canonical.com")),
-		Relation: ofganames.ReaderRelation,
-		Target:   ofganames.ConvertTag(names.NewModelTag("00000002-0000-0000-0000-000000000001")),
-	}},
-}, {
-	name:            "UserNotAuthorized",
-	env:             revokeModelAccessTestEnv,
-	username:        "charlie@canonical.com",
-	uuid:            "00000002-0000-0000-0000-000000000001",
-	targetUsername:  "bob@canonical.com",
-	access:          "write",
-	expectError:     `unauthorized`,
-	expectErrorCode: errors.CodeUnauthorized,
-}, {
-	name:           "DialError",
-	env:            revokeModelAccessTestEnv,
-	dialError:      errors.E("test dial error"),
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "bob@canonical.com",
-	access:         "write",
-	expectError:    `test dial error`,
-}, {
-	name:           "unknown access",
-	env:            revokeModelAccessTestEnv,
-	username:       "alice@canonical.com",
-	uuid:           "00000002-0000-0000-0000-000000000001",
-	targetUsername: "bob@canonical.com",
-	access:         "some-unknown-access",
-	expectError:    `failed to recognize given access: "some-unknown-access"`,
-}}
-
-//nolint:gocognit
-func TestRevokeModelAccess(t *testing.T) {
-	c := qt.New(t)
-
-	for _, t := range revokeModelAccessTests {
-		tt := t
-		c.Run(tt.name, func(c *qt.C) {
-			ctx := context.Background()
-
-			dialer := &jimmtest.Dialer{
-				API: &jimmtest.API{},
-				Err: tt.dialError,
-			}
-			j := jimmtest.NewJIMM(c, &jimm.Parameters{
-				Dialer: dialer,
-			})
-
-			env := jimmtest.ParseEnvironment(c, tt.env)
-			env.PopulateDBAndPermissions(c, j.ResourceTag(), j.Database, j.OpenFGAClient)
-
-			if len(tt.extraInitialTuples) > 0 {
-				err := j.OpenFGAClient.AddRelation(ctx, tt.extraInitialTuples...)
-				c.Assert(err, qt.IsNil)
-			}
-
-			if tt.expectRemovedRelations != nil {
-				for _, tuple := range tt.expectRemovedRelations {
-					value, err := j.OpenFGAClient.CheckRelation(ctx, tuple, false)
-					c.Assert(err, qt.IsNil)
-					c.Assert(value, qt.IsTrue, qt.Commentf("expected the tuple to exist before revoking"))
-				}
-			}
-
-			dbUser := env.User(tt.username).DBObject(c, j.Database)
-			user := openfga.NewUser(&dbUser, j.OpenFGAClient)
-
-			err := j.RevokeModelAccess(ctx, user, names.NewModelTag(tt.uuid), names.NewUserTag(tt.targetUsername), jujuparams.UserAccessPermission(tt.access))
-			c.Assert(dialer.IsClosed(), qt.IsTrue)
-			if tt.expectError != "" {
-				c.Check(err, qt.ErrorMatches, tt.expectError)
-				if tt.expectErrorCode != "" {
-					c.Check(errors.ErrorCode(err), qt.Equals, tt.expectErrorCode)
-				}
-				return
-			}
-			c.Assert(err, qt.IsNil)
-			if tt.expectRemovedRelations != nil {
-				for _, tuple := range tt.expectRemovedRelations {
-					value, err := j.OpenFGAClient.CheckRelation(ctx, tuple, false)
-					c.Assert(err, qt.IsNil)
-					c.Assert(value, qt.IsFalse, qt.Commentf("expected the tuple to be removed after revoking"))
-				}
-			}
-			if tt.expectRelations != nil {
-				for _, tuple := range tt.expectRelations {
-					value, err := j.OpenFGAClient.CheckRelation(ctx, tuple, false)
-					c.Assert(err, qt.IsNil)
-					c.Assert(value, qt.IsTrue, qt.Commentf("expected the tuple to exist after revoking"))
-				}
-			}
 		})
 	}
 }
