+++ internal/jimm/permissions/relations_test.go
@@ -1,10 +1,9 @@
 // Copyright 2024 Canonical.
 
-package jimm_test
+package permissions_test
 
 import (
 	"context"
-	"testing"
 
 	qt "github.com/frankban/quicktest"
 
@@ -16,9 +15,9 @@ import (
 	apiparams "github.com/canonical/jimm/v3/pkg/api/params"
 )
 
-func TestListRelationshipTuples(t *testing.T) {
+func (s *permissionManagerSuite) TestListRelationshipTuples(c *qt.C) {
 	// setup
-	c := qt.New(t)
+	c.Parallel()
 	ctx := context.Background()
 
 	j := jimmtest.NewJIMM(c, nil)
@@ -28,7 +27,7 @@ func TestListRelationshipTuples(t *testing.T) {
 
 	user, _, controller, model, _, _, _, _ := jimmtest.CreateTestControllerEnvironment(ctx, c, j.Database)
 
-	err := j.AddRelation(ctx, u, []apiparams.RelationshipTuple{
+	err := s.manager.AddRelation(ctx, u, []apiparams.RelationshipTuple{
 		{
 			Object:       user.Tag().String(),
 			Relation:     names.ReaderRelation.String(),
@@ -137,7 +136,7 @@ func TestListRelationshipTuples(t *testing.T) {
 
 	for _, t := range testCases {
 		c.Run(t.description, func(c *qt.C) {
-			tuples, _, err := j.ListRelationshipTuples(ctx, u, apiparams.RelationshipTuple{
+			tuples, _, err := s.manager.ListRelationshipTuples(ctx, s.adminUser, apiparams.RelationshipTuple{
 				Object:       t.object,
 				Relation:     t.relation,
 				TargetObject: t.targetObject,
@@ -152,8 +151,8 @@ func TestListRelationshipTuples(t *testing.T) {
 	}
 }
 
-func TestListObjectRelations(t *testing.T) {
-	c := qt.New(t)
+func (s *permissionManagerSuite) TestListObjectRelations(c *qt.C) {
+	c.Parallel()
 	ctx := context.Background()
 
 	j := jimmtest.NewJIMM(c, nil)
@@ -163,7 +162,7 @@ func TestListObjectRelations(t *testing.T) {
 
 	user, group, controller, model, _, cloud, _, _ := jimmtest.CreateTestControllerEnvironment(ctx, c, j.Database)
 
-	err := j.AddRelation(ctx, u, []apiparams.RelationshipTuple{
+	err := s.manager.AddRelation(ctx, u, []apiparams.RelationshipTuple{
 		{
 			Object:       user.Tag().String(),
 			Relation:     names.ReaderRelation.String(),
@@ -250,7 +249,7 @@ func TestListObjectRelations(t *testing.T) {
 			tuples := []openfga.Tuple{}
 			numPages := 0
 			for {
-				res, nextToken, err := j.ListObjectRelations(ctx, u, t.object, t.pageSize, token)
+				res, nextToken, err := s.manager.ListObjectRelations(ctx, s.adminUser, t.object, t.pageSize, token)
 				if t.expectedError != "" {
 					c.Assert(err, qt.ErrorMatches, t.expectedError)
 					break
