+++ internal/jimm/service_account_test.go
@@ -8,14 +8,12 @@ import (
 
 	qt "github.com/frankban/quicktest"
 	jujuparams "github.com/juju/juju/rpc/params"
-	"github.com/juju/names/v5"
 
 	"github.com/canonical/jimm/v3/internal/dbmodel"
 	"github.com/canonical/jimm/v3/internal/jimm"
 	"github.com/canonical/jimm/v3/internal/openfga"
 	ofganames "github.com/canonical/jimm/v3/internal/openfga/names"
 	"github.com/canonical/jimm/v3/internal/testutils/jimmtest"
-	jimmnames "github.com/canonical/jimm/v3/pkg/names"
 )
 
 func TestAddServiceAccount(t *testing.T) {
@@ -152,96 +150,4 @@ func TestCopyServiceAccountCredentialWithMissingCredential(t *testing.T) {
 	}
 	_, _, err = j.CopyServiceAccountCredential(ctx, user, svcAcc, cred.ResourceTag())
 	c.Assert(err, qt.ErrorMatches, "cloudcredential .* not found")
-}
-
-func TestGrantServiceAccountAccess(t *testing.T) {
-	c := qt.New(t)
-
-	tests := []struct {
-		about                     string
-		grantServiceAccountAccess func(ctx context.Context, user *openfga.User, tags []string) error
-		clientID                  string
-		tags                      []string
-		username                  string
-		addGroups                 []string
-		expectedError             string
-	}{{
-		about: "Valid request",
-		grantServiceAccountAccess: func(ctx context.Context, user *openfga.User, tags []string) error {
-			return nil
-		},
-		addGroups: []string{"1"},
-		tags: []string{
-			"user-alice",
-			"user-bob",
-			"group-1#member",
-		},
-		clientID: "fca1f605-736e-4d1f-bcd2-aecc726923be@serviceaccount",
-		username: "alice",
-	}, {
-		about: "Group that doesn't exist",
-		grantServiceAccountAccess: func(ctx context.Context, user *openfga.User, tags []string) error {
-			return nil
-		},
-		tags: []string{
-			"user-alice",
-			"user-bob",
-			// This group doesn't exist.
-			"group-bar",
-		},
-		clientID:      "fca1f605-736e-4d1f-bcd2-aecc726923be@serviceaccount",
-		username:      "alice",
-		expectedError: "group bar not found",
-	}, {
-		about: "Invalid tags",
-		grantServiceAccountAccess: func(ctx context.Context, user *openfga.User, tags []string) error {
-			return nil
-		},
-		tags: []string{
-			"user-alice",
-			"user-bob",
-			"controller-jimm",
-		},
-		clientID:      "fca1f605-736e-4d1f-bcd2-aecc726923be@serviceaccount",
-		username:      "alice",
-		expectedError: "invalid entity - not user or group",
-	}}
-
-	for _, test := range tests {
-		test := test
-		c.Run(test.about, func(c *qt.C) {
-			j := jimmtest.NewJIMM(c, nil)
-
-			var u dbmodel.Identity
-			u.SetTag(names.NewUserTag(test.clientID))
-			svcAccountIdentity := openfga.NewUser(&u, j.OpenFGAClient)
-			svcAccountIdentity.JimmAdmin = true
-			if len(test.addGroups) > 0 {
-				for _, name := range test.addGroups {
-					_, err := j.GroupManager().AddGroup(context.Background(), svcAccountIdentity, name)
-					c.Assert(err, qt.IsNil)
-				}
-			}
-			svcAccountTag := jimmnames.NewServiceAccountTag(test.clientID)
-
-			err := j.GrantServiceAccountAccess(context.Background(), svcAccountIdentity, svcAccountTag, test.tags)
-			if test.expectedError == "" {
-				c.Assert(err, qt.IsNil)
-				for _, tag := range test.tags {
-					parsedTag, err := j.ParseAndValidateTag(context.Background(), tag)
-					c.Assert(err, qt.IsNil)
-					tuple := openfga.Tuple{
-						Object:   parsedTag,
-						Relation: ofganames.AdministratorRelation,
-						Target:   ofganames.ConvertTag(jimmnames.NewServiceAccountTag(test.clientID)),
-					}
-					ok, err := j.OpenFGAClient.CheckRelation(context.Background(), tuple, false)
-					c.Assert(err, qt.IsNil)
-					c.Assert(ok, qt.IsTrue)
-				}
-			} else {
-				c.Assert(err, qt.ErrorMatches, test.expectedError)
-			}
-		})
-	}
 }
